Create ElementAppearingMoreThan25%.cpp
In this question we are asked to find element that appear more than 25%. 

If we see it in different way, it simply boils down to find element which is appearing maximum amount of time. Therefore, by using map we can find the frequency of each element and return the element having the maximum frequency.

Second way, we can even use binary search as the array is sorted.