Create AmountOfTimeForBinTreeToGetInfected.cpp
Intuition :

We will make adjacency list by traversing the graph. The purpose of making this is to traverse back and forth. So simply we will convert it into an undirected graph.

Then just simply apply BFS.

Approach :

We can use unordered map to create a adjacency list.
We will create a queue and visited array for BFS traversal and push the start.
For adjacency list we will insert the value of parent and check for its left and right child.
return minutes - 1 in the end as while doing minutes++ we do it one more than required.

Complexity :
Time complexity: O(N) we traverse twice though.
Space complexity: O(N)