Create FindBottomLeftOfBT.cpp
Intuition
Using Preorder find the left view of binary tree. And return the last element of ans.
To find the left view, do a preorder traversal that means the first node will be the ans. Create a data structure and check is the size of the data structure is same as level(that means new node, insert it)else discard.
return the answer's last element in this question.

Complexity
Time complexity: O(N)
Space complexity: O(H)