Create BinTreeRightSideView.cpp
Intuition
Lets think of preorder traversal, that is, root, left and right. Now if we do it in reverse then we can get root right left. Root will be as it is as its preorder. Therefore, we can say that first node of taversal is our ans.

We will keep a data structure which will store our ans. If the node level is same as the size of data structure. Then store the value else it simply means that it's not the first node and we will not consider it.

Complexity
Time complexity: O(N)
Space complexity: O(H)